Setup and teardown of profiling in Pre/PostRun functions
